// src/api/REST/Transaction/SimpleAggregator/Data/CreateSimpleTransactionPayloadDTO.d.ts
import { SaturnSwapError } from '@/types/Classes/SaturnSwapError';
import { FailTransaction } from '@/types/Transactions/FailTransaction';
import { SuccessTransaction } from '@/types/Transactions/SuccessTransaction';

export interface CreateSimpleTransactionPayloadDTO {
    successTransactions?: SuccessTransaction[];
    failTransactions?: FailTransaction[];
    error?: SaturnSwapError;
}